created comments posts and replies migration model and controller

+++ database/migrations/2024_10_25_041925_create_user_details_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('email')->unique()->index();
             $table->string('country'); 
             $table->date('dob'); 
-            $table->unsignedBigInteger('mobile')->unique()->index(); 
+            $table->unsignedInteger('mobile')->unique()->index(); 
             $table->string('blood_group'); 
             $table->text('address'); 
             $table->string('img_path')->nullable();
